Refactor Ui components.
Update SidebarTab work flow for DG.
