+++ src/main/java/seedu/studybananas/ui/FlashcardUi.java
@@ -13,6 +13,9 @@ import seedu.studybananas.logic.commands.commandresults.FlashcardCommandResult;
 import seedu.studybananas.logic.commands.commandresults.GeneralCommandResult;
 import seedu.studybananas.logic.commands.exceptions.CommandException;
 import seedu.studybananas.logic.parser.exceptions.ParseException;
+import seedu.studybananas.ui.commons.ResultDisplay;
+import seedu.studybananas.ui.flashcardui.FlashcardSetListPanel;
+import seedu.studybananas.ui.flashcardui.FlashcardsDisplay;
 import seedu.studybananas.ui.listeners.CommandResultStateListener;
 import seedu.studybananas.ui.listeners.UiStateListener;
 import seedu.studybananas.ui.util.GlobalState;
