+++ src/main/java/seedu/studybananas/ui/scheduleui/TaskListPanel.java
@@ -1,4 +1,4 @@
-package seedu.studybananas.ui;
+package seedu.studybananas.ui.scheduleui;
 
 import java.util.logging.Logger;
 
@@ -10,6 +10,7 @@ import javafx.scene.layout.Region;
 import seedu.studybananas.commons.core.LogsCenter;
 import seedu.studybananas.logic.Logic;
 import seedu.studybananas.model.task.Task;
+import seedu.studybananas.ui.UiPart;
 
 /**
  * Panel containing the list of persons.
